perf: keep track of the minting NFTs

+++ src/modules/ethEventsScraper/service_layer/ethEventsScraper.service.ts
@@ -94,6 +94,9 @@ export class EthEventsScraperService {
         const collection = await this.nftCollectionRepository.findOne({
           where: { address: event.contract_address.toLowerCase() },
         });
+        let mintingNft = await this.mintingNftRepository.findOne({
+          where: { tokenUri, collectionId: collection.id },
+        });
 
         if (!user || !collection) continue;
 
@@ -103,23 +106,23 @@ export class EthEventsScraperService {
         await this.attachNftEdition(collection, event, nft, editionUUID);
         this.attachNftDataFromEvent(nft, event);
         this.attachNftDataFromTokenUri(nft, response);
-        await this.attachNftNumberOfEditions(nft, tokenUri, collection.id);
+        await this.attachNftNumberOfEditions(nft, tokenUri, collection.id, mintingNft);
 
         event.processed = true;
         await this.createNftEventRepository.save(event);
         await this.nftRepository.save(nft);
+
+        mintingNft.mintedEditions = mintingNft.mintedEditions + 1;
+        mintingNft = await this.mintingNftRepository.save(mintingNft);
+        if (mintingNft.mintedEditions === mintingNft.numberOfEditions) {
+          await this.mintingNftRepository.delete({ tokenUri });
+        }
       }
-      await this.mintingNftRepository.delete({ tokenUri });
     }
   }
 
-  private async attachNftNumberOfEditions(nft: Nft, tokenUri: string, collectionId: number) {
+  private async attachNftNumberOfEditions(nft: Nft, tokenUri: string, collectionId: number, mintingNft: MintingNft) {
     nft.numberOfEditions = 1;
-
-    // This applies only for Saved NFTs. It can be removed, but saves the COUNT query below
-    const mintingNft = await this.mintingNftRepository.findOne({
-      where: { tokenUri, collectionId },
-    });
 
     if (mintingNft) {
       nft.numberOfEditions = mintingNft.numberOfEditions;
